refactor: restructure api folder in file-based requests
-create request, validation and schemas for each file
