+++ src/services/spotifyPlaylistGuardApi/routes/patchPlaylistAllowedUsers.ts
@@ -1,0 +1,58 @@
+import { GetServerSidePropsContext } from 'next';
+import { z } from 'zod';
+import { getToken } from '../auth';
+
+const playlistSchema = z.object({
+    id: z.string(),
+    spotify_id: z.string(),
+    collaborative: z.boolean(),
+    description: z.string(),
+    external_url: z.string(),
+    href: z.string(),
+    name: z.string(),
+    userId: z.string(),
+    public: z.boolean(),
+    snapshot_id: z.string(),
+    uri: z.string(),
+    active: z.boolean(),
+    allowed_userIds: z.array(z.string()),
+    createdAt: z.string(),
+    updatedAt: z.string(),
+});
+
+function validatePlaylistSchema(payload: unknown) {
+    const validation = playlistSchema.safeParse(payload);
+    const { success } = validation;
+
+    return success ? validation.data : undefined;
+}
+
+export async function patchPlaylistAllowedUsers(
+    playlistId: string,
+    userIds: string[],
+    context?: GetServerSidePropsContext,
+) {
+    const token = getToken(context);
+    const apiUrl = process.env.NEXT_PUBLIC_API_URL || '';
+    const response = await fetch(
+        `${apiUrl}/playlists/allowUsers/${playlistId}`,
+        {
+            method: 'PATCH',
+            headers: {
+                Authorization: `Bearer ${token}`,
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({ allowed_userIds: userIds }),
+        },
+    );
+
+    if (response.status !== 200) throw new Error('Invalid response');
+
+    const resBody = await response.json().catch(() => ({}));
+
+    const playlist = validatePlaylistSchema(resBody);
+
+    if (!playlist) throw new Error('Invalid response');
+
+    return;
+}
